Added search functionality
