+++ client/src/App.css
@@ -1,5 +1,5 @@
 html, body {
-  min-height: 100%;
+  height: 100%;
   padding: 0;
   margin: 0;
 }
@@ -201,8 +201,6 @@ h1 {
   }
 }
 
-/* MOVE TO NEW FILE BELOW */
-
 .Queue {
   position: absolute;
   top: 0;
@@ -212,6 +210,9 @@ h1 {
   background-color: #282C34;
   opacity: 0.85;
   color: white;
+
+  display: flex;
+  flex-flow: column;
 }
 
 .Queue-Empty {
@@ -228,6 +229,36 @@ h1 {
   margin-top: 40px;
   margin-bottom: 0px;
   width: 80%;
+}
+
+.Queue-Search-Results {
+  width: 80%;
+  margin: auto;
+  margin-top: 20px;
+  margin-bottom: 50px;
+  flex-grow: 1;
+  overflow: auto;
+}
+
+.Queue-Search-Cell {
+  display: grid;
+  grid-template-columns: 90px auto;
+  border-radius: 12px;
+  background-color: #17191c;
+  margin-top: 15px;
+}
+
+.Queue-Search-Image {
+  height: 85px;
+  width: 85px;
+  padding: 10px;
+  border-radius: 15px;
+}
+
+.Queue-Search-Text {
+  vertical-align: middle;
+  padding: 10px;
+  margin: auto;
 }
 
 .Queue-Body {
@@ -258,7 +289,6 @@ textarea {
   font-size: 16px;
 }
 
-/* NEW FILE */
 .Param-Grid {
   display: grid;
   grid-gap: 30px;
